convert jsx to tsx
