+++ src/components/CaseCard.tsx
