+++ src/components/FooterBar.tsx
