+++ src/components/MainContent.tsx
