+++ src/index.tsx
@@ -1,0 +1,171 @@
+import React, { useEffect, useState } from "react";
+import i18n from "./i18n";
+import "bootstrap/dist/css/bootstrap.min.css";
+import ReactDOM from "react-dom/client";
+import { I18nextProvider } from "react-i18next";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  useNavigate,
+  useParams,
+} from "react-router-dom";
+import RecipeCategory from "./Pages/RecipeCategory";
+import RecipeDetail from "./Pages/RecipeDetail";
+import AddRecipe from "./Pages/AddRecipe";
+import HomePage from "./Pages/HomePage";
+import "./styles.css";
+import { CircularProgress, Box } from "@mui/material";
+import { Provider } from "react-redux";
+import * as storage from "./utils/storage";
+import store from "./store/store";
+
+const rootElement = document.getElementById("root") as HTMLElement;
+const root = ReactDOM.createRoot(rootElement);
+
+function App() {
+  const [recipes, setRecipes] = useState<any>(null);
+  const [selected, setSelected] = useState<any>(null);
+  const [selectedRecipe, setSelectedRecipe] = useState<any>(null);
+  const [selectedCategory, setSelectedCategory] = useState<any>(null);
+  const [loading, setLoading] = useState(true);
+
+  const navigate = useNavigate();
+  const params = useParams();
+
+  useEffect(() => {
+    const fetchData = async () => {
+      setLoading(true);
+      const data = await storage.loadData(false);
+      setRecipes(data);
+
+      const categoryParam = params.category;
+      const titleParam = params.title;
+
+      if (data?.site?.pages && data.site.pages.length > 0) {
+        let initialCategory = data.site.pages[0];
+        if (categoryParam) {
+          const foundCat = data.site.pages.find(
+            (cat: any) => encodeURIComponent(cat.category) === categoryParam
+          );
+          if (foundCat) initialCategory = foundCat;
+        }
+        setSelectedCategory(initialCategory);
+
+        if (titleParam && initialCategory.itemPage) {
+          const foundRecipe = initialCategory.itemPage.find(
+            (rec: any) => encodeURIComponent(rec.title) === titleParam
+          );
+          if (foundRecipe) setSelectedRecipe(foundRecipe);
+        }
+      }
+
+      setLoading(false);
+    };
+    fetchData();
+  }, [params.category, params.title]);
+
+  useEffect(() => {
+    document.body.dir =
+      i18n.language === "he" || i18n.language === "ar" ? "rtl" : "ltr";
+  }, [i18n.language]);
+
+  useEffect(() => {
+    if (selectedRecipe && selectedRecipe.title && selectedRecipe.category) {
+      navigate(
+        `/recipes/${encodeURIComponent(selectedRecipe.category)}/${encodeURIComponent(selectedRecipe.title)}`
+      );
+    }
+  }, [selectedRecipe, navigate]);
+
+  return (
+    <>
+      {loading && (
+        <Box
+          sx={{
+            width: "100vw",
+            height: "100vh",
+            display: "flex",
+            alignItems: "center",
+            justifyContent: "center",
+            position: "fixed",
+            left: 0,
+            top: 0,
+            zIndex: 2000,
+            background: "#fffce8",
+          }}
+        >
+          <CircularProgress size={64} />
+        </Box>
+      )}
+      {!loading && recipes && (
+        <Routes>
+          <Route
+            path="/"
+            element={
+              <HomePage
+                recipes={recipes}
+                setRecipes={setRecipes}
+                selectedRecipe={selectedRecipe}
+                setSelectedRecipe={setSelectedRecipe}
+                setSelectedCategory={setSelectedCategory}
+                selectedCategory={selectedCategory}
+              />
+            }
+          />
+          <Route
+            path="/recipes/:category"
+            element={
+              <RecipeCategory
+                recipes={recipes}
+                setRecipes={setRecipes}
+                selectedRecipe={selectedRecipe}
+                setSelectedRecipe={setSelectedRecipe}
+                setSelectedCategory={setSelectedCategory}
+                selectedCategory={selectedCategory}
+              />
+            }
+          />
+          <Route
+            path="/recipes/:category/:title"
+            element={
+              <RecipeDetail
+                recipes={recipes}
+                setSelectedCategory={setSelectedCategory}
+                selectedCategory={selectedCategory}
+                setSelected={setSelected}
+                selectedRecipe={selectedRecipe}
+                setSelectedRecipe={setSelectedRecipe}
+              />
+            }
+          />
+          <Route
+            path="/recipes/:category/add"
+            element={
+              <AddRecipe
+                recipes={recipes}
+                setRecipes={setRecipes}
+                setSelectedCategory={setSelectedCategory}
+                selectedCategory={selectedCategory}
+                selectedRecipe={selectedRecipe}
+                setSelectedRecipe={setSelectedRecipe}
+              />
+            }
+          />
+        </Routes>
+      )}
+    </>
+  );
+}
+
+root.render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <I18nextProvider i18n={i18n}>
+        <Router>
+          <App />
+        </Router>
+      </I18nextProvider>
+    </Provider>
+  </React.StrictMode>
+);
